healthtwin/requirements.txt
# HealthTwin — AI-Powered Digital Twin for Personal Health
# Python 3.11 | Hackathon Build
# Core data
pandas>=1.5.0,<2.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0

# Feature engineering
tsfresh>=0.20.0

# Physiological signal processing
neurokit2>=0.2.7
pyEDFlib>=0.1.36

# Machine learning
scikit-learn>=1.3.0
xgboost<2.1.0
shap>=0.47.2
imbalanced-learn>=0.12.0

# Explainability
shap>=0.43.0
dice-ml>=0.11

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Dashboard
streamlit>=1.30.0

# LLM integration (GPT-4o-mini)
openai>=1.30.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
PyYAML>=6.0.1
loguru>=0.7.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
